docker_tasks/build_stac/requirements.txt
aws-lambda-powertools
awslambdaric
boto3
pystac==1.4.0
python-cmr
rasterio==1.3.0
rio-stac==0.4.2
shapely
smart-open==6.3.0
pydantic==1.9.1
